[ag] #149 - tests - types samples
